Centralize setup SQL script
